requirements.txt
python_telegram_bot==13.3
numpy==1.18.5
Werkzeug==1.0.1
opencv_python==4.4.0.44
openpyxl==3.0.5
Flask==1.1.2
matplotlib==3.3.2
colour==0.1.5
Pillow==9.0.1
telegram==0.0.1
